@@ -11,7 +11,7 @@
   "require": {
     "anhskohbo/no-captcha": "^3.2",
     "chaseconey/laravel-datadog-helper": ">=1.2.0",
-    "doctrine/dbal": "^2.0",
+    "doctrine/dbal": "^3.1",
     "egulias/email-validator": "*",
     "elasticsearch/elasticsearch": "^7.12.0",
     "ezyang/htmlpurifier": "*",
@@ -21,8 +21,9 @@
     "itsgoingd/clockwork": "dev-master#fd0e893",
     "jenssegers/agent": "*",
     "knuckleswtf/scribe": "^2.7",
-    "laravel/framework": "^6.20.14",
+    "laravel/framework": "^8.44",
     "laravel/helpers": "^1.1",
+    "laravel/legacy-factories": "^1.1",
     "laravel/passport": "*",
     "laravel/slack-notification-channel": "^2.0",
     "laravel/tinker": "*",
@@ -41,12 +42,11 @@
   "require-dev": {
     "beyondcode/laravel-query-detector": "*",
     "dealerdirect/phpcodesniffer-composer-installer": "^0.7.0",
-    "dms/phpunit-arraysubset-asserts": "^0.2.1",
-    "facade/ignition": "^1.0",
-    "fzaninotto/faker": "*",
+    "dms/phpunit-arraysubset-asserts": "^0.3.0",
+    "facade/ignition": "^2.10",
     "laravel/dusk": "*",
     "mockery/mockery": "*",
-    "nunomaduro/collision": "^3.0",
+    "nunomaduro/collision": "^5.4",
     "phpunit/phpunit": ">=9.5",
     "slevomat/coding-standard": "^6.4",
     "squizlabs/php_codesniffer": "^3.5",
@@ -58,7 +58,8 @@
       "database"
     ],
     "psr-4": {
-      "App\\": "app/"
+      "App\\": "app/",
+      "Database\\Seeders\\": "database/seeders/"
     },
     "files": [
       "app/helpers.php"
@@ -86,14 +87,12 @@
     "sort-packages": true,
     "optimize-autoloader": true,
     "platform": {
-      "php": "7.4.0"
+      "php": "8.0.0"
     }
   },
   "extra": {
     "laravel": {
-      "dont-discover": [
-        "laravel/passport"
-      ]
+      "dont-discover": []
     }
   }
 }
