@@ -15,7 +15,7 @@ class SessionManager extends \Illuminate\Session\SessionManager
      */
     protected function buildSession($handler)
     {
-        return new Store($this->app['config']['session.cookie'], $handler);
+        return new Store($this->config->get('session.cookie'), $handler);
     }
 
     // copied from upstream but with custom CacheBasedSessionHandler
