@@ -3,6 +3,9 @@
 // Copyright (c) ppy Pty Ltd <contact@ppy.sh>. Licensed under the GNU Affero General Public License v3.0.
 // See the LICENCE file in the repository root for full licence text.
 
+namespace Database\Seeders\ModelSeeders;
+
+use App\Models\BanchoStats;
 use Carbon\Carbon;
 use Illuminate\Database\Seeder;
 
@@ -18,7 +21,7 @@ class BanchoStatsSeeder extends Seeder
         $date = new Carbon();
 
         //Create 500 new data points
-        factory(App\Models\BanchoStats::class, 500)->make()->each(function ($stat) use ($date) {
+        factory(BanchoStats::class, 500)->make()->each(function ($stat) use ($date) {
             $stat->date = $date;
             $stat->save();
 
