@@ -3,7 +3,13 @@
 // Copyright (c) ppy Pty Ltd <contact@ppy.sh>. Licensed under the GNU Affero General Public License v3.0.
 // See the LICENCE file in the repository root for full licence text.
 
+namespace Database\Seeders\ModelSeeders;
+
+use App\Models\Beatmapset;
+use App\Models\Forum;
 use App\Models\User;
+use Exception;
+use Faker\Factory as Faker;
 use Illuminate\Database\Seeder;
 
 class ForumSeeder extends Seeder
@@ -15,45 +21,45 @@ class ForumSeeder extends Seeder
      */
     public function run()
     {
-        $faker = Faker\Factory::create();
+        $faker = Faker::create();
 
         try {
             // Create appropriate forum permissions
             $authOptionIds = [];
 
             foreach (['f_post', 'f_postcount', 'f_read', 'f_reply'] as $authOption) {
-                $option = new App\Models\Forum\AuthOption();
+                $option = new Forum\AuthOption();
                 $option->auth_option = $authOption;
                 $option->save();
 
                 $authOptionIds[] = $option->auth_option_id;
             }
 
-            $beatmapCount = App\Models\Beatmapset::count();
+            $beatmapCount = Beatmapset::count();
             if ($beatmapCount > 0) {
                 // Create beatmap threads
-                $f = factory(App\Models\Forum\Forum::class)->states('parent')->create([
+                $f = factory(Forum\Forum::class)->states('parent')->create([
                     'forum_name' => 'Beatmap Threads',
                     'forum_desc' => 'Beatmap thread info for beatmaps',
                 ]);
 
-                $f2 = $f->subforums()->save(factory(App\Models\Forum\Forum::class)->states('child')->make([
+                $f2 = $f->subforums()->save(factory(Forum\Forum::class)->states('child')->make([
                     'parent_id' => 1,
                     'forum_name' => 'Beatmap Threads',
                     'forum_desc' => 'Beatmap thread info for beatmaps',
                 ]));
 
-                $bms = App\Models\Beatmapset::all();
+                $bms = Beatmapset::all();
                 foreach ($bms as $set) {
                     $user = User::find($set->user_id);
-                    $t = $f2->topics()->save(factory(App\Models\Forum\Topic::class)->make([
+                    $t = $f2->topics()->save(factory(Forum\Topic::class)->make([
                         'forum_id' => $f2->forum_id,
                         'topic_first_poster_name' => $user->username,
                         'topic_poster' => $user->getKey(),
                         'topic_title' => $set->artist.' - '.$set->title,
                     ]));
 
-                    $p = $t->posts()->save(factory(App\Models\Forum\Post::class)->make([
+                    $p = $t->posts()->save(factory(Forum\Post::class)->make([
                         'forum_id' => $f2->forum_id,
                         'poster_id' => $set->user_id,
                         'post_username' => $set->creator,
@@ -69,26 +75,26 @@ class ForumSeeder extends Seeder
             }
 
             // Create userpage threads
-            $f = factory(App\Models\Forum\Forum::class)->states('parent')->create([
+            $f = factory(Forum\Forum::class)->states('parent')->create([
                 'forum_name' => 'User Pages',
                 'forum_desc' => 'Your user profile pages go here!',
             ]);
 
-            $f->subforums()->save(factory(App\Models\Forum\Forum::class)->states('child')->make([
+            $f->subforums()->save(factory(Forum\Forum::class)->states('child')->make([
                 'forum_id' => config('osu.user.user_page_forum_id'),
                 'parent_id' => $f->forum_id,
                 'forum_name' => 'User Pages',
             ]));
 
             // Create 3 forums
-            factory(App\Models\Forum\Forum::class, 3)->states('parent')->create()->each(function ($f) {
+            factory(Forum\Forum::class, 3)->states('parent')->create()->each(function ($f) {
                 for ($i = 0; $i < 4; $i++) {
                     // Subforums for each forum.
-                    $f2 = $f->subforums()->save(factory(App\Models\Forum\Forum::class)->states('child')->make());
+                    $f2 = $f->subforums()->save(factory(Forum\Forum::class)->states('child')->make());
                     // Topics for each subforum
                     for ($j = 0; $j < 3; $j++) {
                         $topicUser = User::orderByRaw('RAND()')->first();
-                        $t = $f2->topics()->save(factory(App\Models\Forum\Topic::class)->make([
+                        $t = $f2->topics()->save(factory(Forum\Topic::class)->make([
                             'forum_id' => $f2->forum_id,
                             'topic_first_poster_name' => $topicUser->username,
                             'topic_poster' => $topicUser->getKey(),
@@ -96,7 +102,7 @@ class ForumSeeder extends Seeder
                         // Replies to the topic
                         for ($k = 0; $k < 5; $k++) {
                             $postUser = User::orderByRaw('RAND()')->first();
-                            $p = $t->posts()->save(factory(App\Models\Forum\Post::class)->make([
+                            $p = $t->posts()->save(factory(Forum\Post::class)->make([
                                 'forum_id' => $f2->forum_id,
                                 'post_username' => $postUser->username,
                                 'poster_id' => $postUser->getKey(),
@@ -110,9 +116,9 @@ class ForumSeeder extends Seeder
                 }
             });
 
-            foreach (App\Models\Forum\Forum::all() as $forum) {
+            foreach (Forum\Forum::all() as $forum) {
                 foreach ($authOptionIds as $optionId) {
-                    $group = new App\Models\Forum\Authorize();
+                    $group = new Forum\Authorize();
 
                     $group->group_id = app('groups')->byIdentifier('default')->getKey();
                     $group->forum_id = $forum->forum_id;
@@ -122,10 +128,8 @@ class ForumSeeder extends Seeder
                     $group->save();
                 }
             }
-        } catch (\Illuminate\Database\QueryException $e) {
-            echo $e->getMessage()."\r\n";
         } catch (Exception $ex) {
-            echo $ex->getMessage()."\r\n";
+            echo $ex->getMessage().PHP_EOL;
         }
     }
 }
