@@ -3,7 +3,14 @@
 // Copyright (c) ppy Pty Ltd <contact@ppy.sh>. Licensed under the GNU Affero General Public License v3.0.
 // See the LICENCE file in the repository root for full licence text.
 
-use Illuminate\Database\QueryException;
+namespace Database\Seeders\ModelSeeders;
+
+use App\Models\BeatmapLeader;
+use App\Models\BeatmapPlaycount;
+use App\Models\FavouriteBeatmapset;
+use App\Models\User;
+use DB;
+use Exception;
 use Illuminate\Database\Seeder;
 
 class UserProfileSeeder extends Seeder
@@ -18,7 +25,7 @@ class UserProfileSeeder extends Seeder
         try {
             // FAVOURITE BEATMAPS AND BEATMAP PLAYCOUNTS FOR EACH USER
 
-            foreach (App\Models\User::all() as $usr) {
+            foreach (User::all() as $usr) {
                 $bms = $usr->scoresBestOsu()->get();
                 if (count($bms) < 1) {
                     $this->command->info('Can\'t seed favourite maps, map playcounts or leaders due to having no beatmap data.');
@@ -30,7 +37,7 @@ class UserProfileSeeder extends Seeder
                 foreach ($bms as $bm) {
                     $beatmapset = $bm->beatmap->beatmapset;
                     DB::table('osu_favouritemaps')->where('user_id', $usr_id)->where('beatmapset_id', $beatmapset->beatmapset_id)->delete();
-                    $fav = new App\Models\FavouriteBeatmapset();
+                    $fav = new FavouriteBeatmapset();
                     $fav->beatmapset_id = $beatmapset->beatmapset_id;
                     $fav->user_id = $usr_id;
                     $fav->save();
@@ -39,7 +46,7 @@ class UserProfileSeeder extends Seeder
 
                     $bm = $bms[rand(0, count($bms) - 1)];
                     DB::table('osu_user_beatmap_playcount')->where('user_id', $usr_id)->where('beatmap_id', $bm['beatmap_id'])->delete();
-                    $playcount = new App\Models\BeatmapPlaycount();
+                    $playcount = new BeatmapPlaycount();
 
                     $playcount->user_id = $usr_id;
                     $playcount->beatmap_id = $bm['beatmap_id'];
@@ -54,17 +61,15 @@ class UserProfileSeeder extends Seeder
                             DB::table('osu_leaders')->where('beatmap_id', $bm['beatmap_id'])->delete();
                         }
                     }
-                    $leader = new App\Models\BeatmapLeader\Osu();
+                    $leader = new BeatmapLeader\Osu();
                     $leader->beatmap_id = $bm['beatmap_id'];
                     $leader->user_id = $usr_id;
                     $leader->score_id = $bm['score_id'];
                     $leader->save();
                 }
             }
-        } catch (QueryException $e) {
-            $this->command->error("Error: Unable to save User Profile Data\r\n".$e->getMessage());
         } catch (Exception $ex) {
-            $this->command->error("Error: Unable to save User Profile Data\r\n".$ex->getMessage());
+            $this->command->error('Error: Unable to save User Profile Data'.PHP_EOL.$ex->getMessage());
         }
     }
 }
