@@ -4,12 +4,12 @@
 --}}
 @php
     // this is pretty much a php conversion of beatmap-discussions/user-card.coffee
-    $topClasses = $bn = 'beatmap-discussion-user-card';
-    $group = optional($user->userGroupsForBadges()->first())->group;
+    $bn = 'beatmap-discussion-user-card';
+    $userGroup = $user->userGroupsForBadges()->first();
     $hideStripe = $hideStripe ?? false;
 @endphp
 
-<div class="{{$topClasses}}" style="{!! css_group_colour($group) !!}">
+<div class="{{$bn}}" style="{!! css_group_colour(optional($userGroup)->group) !!}">
     <div class="{{$bn}}__avatar">
         @if (isset($user))
             <a class="{{$bn}}__user-link" href="{{route('users.show', $user)}}">
@@ -39,13 +39,8 @@
             @endif
         </div>
         <div class="{{$bn}}__user-badge">
-            @if ($group !== null)
-                <div
-                    class="user-group-badge"
-                    title="{{ $group->group_name }}"
-                    data-label="{{ $group->short_name }}"
-                    style="{!! css_group_colour($group) !!}"
-                ></div>
+            @if ($userGroup !== null)
+                @include('objects._user_group_badge', compact('userGroup'))
             @endif
         </div>
     </div>
