@@ -2,6 +2,9 @@
     Copyright (c) ppy Pty Ltd <contact@ppy.sh>. Licensed under the GNU Affero General Public License v3.0.
     See the LICENCE file in the repository root for full licence text.
 --}}
+@php
+    $currentLocaleMeta = current_locale_meta();
+@endphp
 @foreach (nav_links() as $section => $links)
     <div class="navbar-mobile-item">
         <a
@@ -53,16 +56,19 @@
 
         <span class="navbar-mobile-item__locale-flag">
             @include('objects._flag_country', [
-                'countryCode' => locale_flag(App::getLocale()),
+                'countryCode' => $currentLocaleMeta->flag(),
                 'modifiers' => ['small', 'flat'],
             ])
         </span>
 
-        {{ locale_name(App::getLocale()) }}
+        {{ $currentLocaleMeta->name() }}
     </button>
 
     <ul class="navbar-mobile-item__submenu js-click-menu" data-click-menu-id="nav-mobile-locale">
         @foreach (config('app.available_locales') as $locale)
+            @php
+                $localeMeta = locale_meta($locale);
+            @endphp
             <li>
                 <a
                     class="navbar-mobile-item__submenu-item js-click-menu--close"
@@ -72,12 +78,12 @@
                 >
                     <span class="navbar-mobile-item__locale-flag">
                         @include('objects._flag_country', [
-                            'countryCode' => locale_flag($locale),
+                            'countryCode' => $localeMeta->flag(),
                             'modifiers' => ['small', 'flat'],
                         ])
                     </span>
 
-                    {{ locale_name($locale) }}
+                    {{ $localeMeta->name() }}
                 </a>
             </li>
         @endforeach
