@@ -8,8 +8,8 @@ namespace Tests\Controllers\Passport;
 use App\Http\Controllers\Passport\AuthorizationController;
 use App\Models\OAuth\Client;
 use Mockery;
+use Nyholm\Psr7\Factory\Psr17Factory;
 use Tests\TestCase;
-use Zend\Diactoros\ServerRequest;
 
 class AuthorizationControllerTest extends TestCase
 {
@@ -19,20 +19,9 @@ class AuthorizationControllerTest extends TestCase
     {
         $client = factory(Client::class)->create();
 
-        // phpcs:disable
-        $request = new ServerRequest(
-            /* $serverParams */ [],
-            /* $uploadedFiles */ [],
-            /* $uri */ config('app.url').'/oauth/authorize',
-            /* $method */ 'GET',
-            /* $body */ 'php://input',
-            /* $headers */  [],
-            /* $cookies */  [],
-            /* $queryParams */ ['client_id' => $client->getKey(), 'scope' => 'one two three'],
-            /* $parsedBody */ null,
-            /* $protocol*/ '1.1'
-        );
-        // phpcs:enable
+        $request = (new Psr17Factory())
+            ->createServerRequest('GET', config('app.url').'/oauth/authorize')
+            ->withQueryParams(['client_id' => $client->getKey(), 'scope' => 'one two three']);
 
         $actual = $this->invokeMethod($this->controller, 'normalizeRequestScopes', [$request])->getQueryParams()['scope'];
 
